+++ src/main/java/it/polpetta/libris/util/SSL/ISSLContextObtainer.java
@@ -1,4 +1,4 @@
-package it.polpetta.libris.utils.SSL;
+package it.polpetta.libris.util.SSL;
 
 import javax.net.ssl.SSLContext;
